# Compute hashes correctly when constructing internalised collections