@ehildenb 2024-04-03
We must have a testing harness for this, and multiple tests! Ideally unit and integration!