@geo2a 2024-04-03
Yeah, I'm adding unit tests now.