@geo2a 2024-04-04
KEVM performance at commit https://github.com/runtimeverification/hs-backend-booster/commit/3d58ba36c7870f9156f4cd2964f58add73c8133b with 1 worker

| Test                                     | georgy-fix-collection-hashes time | master-bafee82 time | (georgy-fix-collection-hashes/master-bafee82) time
|------------------------------------------|-----------------------------------|---------------------|----------------------------------------------------
| mcd/flopper-file-pass-rough-spec.k       | 439.37                            | 463.1               | 0.9487583675232131
| erc20/hkg/transferFrom-success-1-spec.k  | 64.09                             | 66.82               | 0.9591439688715955
| erc20/ds/transferFrom-failure-2-c-spec.k | 57.93                             | 55.79               | 1.038358128696899
| erc20/ds/transferFrom-success-1-spec.k   | 98.77                             | 94.32               | 1.0471798134011874
| TOTAL                                    | 660.16                            | 680.03              | 0.970780700851433